% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcrypto.R
\name{rcrypto_int}
\alias{rcrypto_int}
\title{Generate random R integers in the range [0, 1] using the 
platform-specific cryptographically secure pseudorandom number generator}
\usage{
rcrypto_int(n)
}
\arguments{
\item{n}{Number of random integers to generate.
Note: if the entropy pool is exhausted on your
system it may not be able to provide the requested number of bytes -
in this case an error is thrown.}
}
\value{
An integer vector.  Note: this function may return \code{NA} values
        (with a very low probability).  Users are advised to use \code{na.omit()}
        (or similar) to remove \code{NA} values if not needed.
}
\description{
Generate random R integers in the range [0, 1] using the 
platform-specific cryptographically secure pseudorandom number generator
}
\examples{
rcrypto_int(16)
}
