% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcrypto.R
\name{rcrypto_unif}
\alias{rcrypto_unif}
\title{Generate uniform random floating point numbers in the range [0, 1] using the 
platform-specific cryptographically secure pseudorandom number generator}
\usage{
rcrypto_unif(n)
}
\arguments{
\item{n}{Number of random values to generate.
Note: if the entropy pool is exhausted on your
system it may not be able to provide the requested number of bytes -
in this case an error is thrown.}
}
\value{
A numeric vector of double precision floating point values in the 
        range [0, 1].  Note: this function never returns \code{NA} values.
}
\description{
Generate uniform random floating point numbers in the range [0, 1] using the 
platform-specific cryptographically secure pseudorandom number generator
}
\examples{
rcrypto_unif(16)
}
